Change background image code to resources.Get
https://discourse.gohugo.io/t/what-is-the-hugo-way-to-give-an-html-section-a-background-image-inline-styling-with-a-resource-get/39727/2
Before this article, what I tried is changing the whole
baseof.html file. The problem of this approach was that
this not only changed the background of the homepage, but also
all other pages as well. Thus, I was trying to look for a solution,
and this article worked out perfectly.

Signed-off-by: Soobin Rho <soobinrho@gmail.com>